refactor: move HospitalRun to root directory
