+++ src/HospitalRun.tsx
@@ -2,15 +2,15 @@ import React from 'react'
 import { Switch, Route } from 'react-router-dom'
 import { useSelector } from 'react-redux'
 import { Toaster } from '@hospitalrun/components'
-import Sidebar from '../components/Sidebar'
-import Permissions from '../model/Permissions'
-import Dashboard from './Dashboard'
-import Patients from '../patients/list/Patients'
-import NewPatient from '../patients/new/NewPatient'
-import ViewPatient from '../patients/view/ViewPatient'
-import { RootState } from '../store'
-import Navbar from '../components/Navbar'
-import PrivateRoute from '../components/PrivateRoute'
+import Sidebar from './components/Sidebar'
+import Permissions from './model/Permissions'
+import Dashboard from './dashboard/Dashboard'
+import Patients from './patients/list/Patients'
+import NewPatient from './patients/new/NewPatient'
+import ViewPatient from './patients/view/ViewPatient'
+import { RootState } from './store'
+import Navbar from './components/Navbar'
+import PrivateRoute from './components/PrivateRoute'
 
 const HospitalRun = () => {
   const { title } = useSelector((state: RootState) => state.title)
