+++ src/__tests__/HospitalRun.test.tsx
@@ -1,4 +1,4 @@
-import '../../__mocks__/matchMediaMock'
+import '../__mocks__/matchMediaMock'
 import React from 'react'
 import { mount } from 'enzyme'
 import { MemoryRouter } from 'react-router'
@@ -7,12 +7,12 @@ import { mocked } from 'ts-jest/utils'
 import thunk from 'redux-thunk'
 import configureMockStore from 'redux-mock-store'
 import { Toaster } from '@hospitalrun/components'
-import NewPatient from '../../patients/new/NewPatient'
-import ViewPatient from '../../patients/view/ViewPatient'
-import PatientRepository from '../../clients/db/PatientRepository'
-import Patient from '../../model/Patient'
-import HospitalRun from '../../containers/HospitalRun'
-import Permissions from '../../model/Permissions'
+import NewPatient from '../patients/new/NewPatient'
+import ViewPatient from '../patients/view/ViewPatient'
+import PatientRepository from '../clients/db/PatientRepository'
+import Patient from '../model/Patient'
+import HospitalRun from '../HospitalRun'
+import Permissions from '../model/Permissions'
 
 const mockStore = configureMockStore([thunk])
